requirements.txt
colorama==0.4.6
numpy==1.24.0
Pillow==9.4.0
pygame==2.1.2
qrcode==7.3.1
termcolor==2.2.0
